Add auto adapter (#6)
This PR moves `BusAdapter` to its own file, and adds `AutoBusAdapter`
which uses `IntoAddress` and `FromAddress` to implement the address
translation. Error mapping has been moved for both to using the `From`
trait. I'm hoping they will make it possible for the compiler to
optimize better than with closures, as well as being easier to use since
the translation only has to be defined once, but it also takes away some
fine-grained control.

It also includes some clean up